requirements.txt
asgiref==3.7.2
astroid==2.15.6
colorama==0.4.6
dill==0.3.6
distlib==0.3.6
Django==4.2.3
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
filelock==3.12.2
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
phonenumbers==8.13.16
platformdirs==3.9.1
pylint==2.17.4
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
typing-extensions==4.7.1
tzdata==2023.3
virtualenv==20.24.0
virtualenvwrapper-win==1.2.7
wrapt==1.15.0
